tgg/requirements.txt
pytelegrambotapi
fastapi
jinja2
sqlalchemy
pydantic-settings
psycopg2-binary
odoorpc
pandas
pyproj==3.3
apscheduler
uvicorn
fastapi-security
python-multipart
python-jose
passlib
itsdangerous
sqlalchemyseed
alembic
